+++ src/modules/headset/headset.h
@@ -29,6 +29,7 @@ typedef struct {
   bool mask;
   bool stencil;
   bool antialias;
+  bool hdr;
   bool submitDepth;
   bool overlay;
   uint32_t overlayOrder;
@@ -199,6 +200,7 @@ float lovrHeadsetGetRefreshRate(void);
 bool lovrHeadsetSetRefreshRate(float refreshRate);
 void lovrHeadsetGetFoveation(FoveationLevel* level, bool* dynamic);
 bool lovrHeadsetSetFoveation(FoveationLevel level, bool dynamic);
+bool lovrHeadsetIsHDR(void);
 bool lovrHeadsetIsPassthroughSupported(PassthroughMode mode);
 PassthroughMode lovrHeadsetGetPassthrough(void);
 bool lovrHeadsetSetPassthrough(PassthroughMode mode);